py3-pdp/pyproject.toml
[tool.poetry]
name = "py3-pdp"
version = "0.1.0"
description = ""
authors = ["Hubert Bugaj <lesny.rumcajs+github@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyzmq = "^26.4.0"
python-dotenv = "^1.1.0"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
